+++ src/backend/InvenTree/InvenTree/tests.py
@@ -4,7 +4,6 @@ import os
 import time
 from datetime import datetime, timedelta
 from decimal import Decimal
-from pathlib import Path
 from unittest import mock
 from zoneinfo import ZoneInfo
 
@@ -1266,22 +1265,6 @@ class TestSettings(InvenTreeTestCase):
                 str(test_file).lower(), str(config.get_config_file()).lower()
             )
 
-        # LEGACY - old path
-        if settings.DOCKER:  # pragma: no cover
-            # In Docker, the legacy path is not used
-            return
-        legacy_path = config.get_base_dir().joinpath('config.yaml')
-        assert not legacy_path.exists(), (
-            'Legacy config file does exist, stopping as a percaution!'
-        )
-        self.assertTrue(test_file.exists(), f'Test file {test_file} does not exist!')
-        test_file.rename(legacy_path)
-        self.assertIn(
-            'src/backend/inventree/config.yaml', str(config.get_config_file()).lower()
-        )
-        # Clean up again
-        legacy_path.unlink(missing_ok=True)
-
     def test_helpers_plugin_file(self):
         """Test get_plugin_file."""
         # normal run - not configured
@@ -1312,22 +1295,6 @@ class TestSettings(InvenTreeTestCase):
         with in_env_context({'INVENTREE_SECRET_KEY_FILE': str(test_file)}):
             self.assertIn(str(test_file), str(config.get_secret_key(return_path=True)))
 
-        # LEGACY - old path
-        if settings.DOCKER:  # pragma: no cover
-            # In Docker, the legacy path is not used
-            return
-        legacy_path = config.get_base_dir().joinpath('secret_key.txt')
-        assert not legacy_path.exists(), (
-            'Legacy secret key file does exist, stopping as a percaution!'
-        )
-        test_file.rename(legacy_path)
-        self.assertIn(
-            'src/backend/inventree/secret_key.txt',
-            str(config.get_secret_key(return_path=True)).lower(),
-        )
-        # Clean up again
-        legacy_path.unlink(missing_ok=True)
-
         # Test with content set per environment
         with in_env_context({'INVENTREE_SECRET_KEY': '123abc123'}):
             self.assertEqual(config.get_secret_key(), '123abc123')
@@ -1352,27 +1319,6 @@ class TestSettings(InvenTreeTestCase):
         # Override with environment variable
         with in_env_context({'INVENTREE_OIDC_PRIVATE_KEY': '123abc123'}):
             self.assertEqual(config.get_oidc_private_key(), '123abc123')
-
-        # LEGACY - old path
-        if settings.DOCKER:  # pragma: no cover
-            # In Docker, the legacy path is not used
-            return
-        legacy_path = config.get_base_dir().joinpath('oidc.pem')
-        assert not legacy_path.exists(), (
-            'Legacy OIDC private key file does exist, stopping as a precaution!'
-        )
-        test_file.rename(legacy_path)
-        assert isinstance(trgt_path, Path)
-        new_path = trgt_path.rename(
-            trgt_path.parent / '_oidc.pem'
-        )  # move out current config
-        self.assertIn(
-            'src/backend/inventree/oidc.pem',
-            str(config.get_oidc_private_key(return_path=True)).lower(),
-        )
-        # Clean up again
-        legacy_path.unlink(missing_ok=True)
-        new_path.rename(trgt_path)  # restore original path for current config
 
     def test_helpers_setting(self):
         """Test get_setting."""