+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Breaking Changes - Remove deprecated features and APIs
 
+- [#9814](https://github.com/inventree/InvenTree/pull/9814) - removes legacy config path support
 
 ## Unreleased - YYYY-MM-DD